Add swiftlint for ci

+++ ci/swift/Sources/AllTests/Tests/CiFoundation/FunctionalTests/FileDownloaderTests.swift
@@ -5,14 +5,22 @@ import Bash
 
 public final class FileDownloaderTests: XCTestCase {
     func test() {
-        let di = TeamcityBuildDi()
-        try! di.bootstrap()
-        let fileDownloader: FileDownloader = try! di.resolve()
-        
-        let path = try! fileDownloader.download(url: URL(string: "https://example.com")!)
-        
-        XCTAssert(FileManager.default.fileExists(atPath: path))
-        
-        try? FileManager.default.removeItem(atPath: path)
+        do {
+            let di = TeamcityBuildDi()
+            try di.bootstrap()
+            let fileDownloader: FileDownloader = try di.resolve()
+            
+            guard let url = URL(string: "https://example.com") else {
+                throw ErrorString("Couldn't make url")
+            }
+            
+            let path = try fileDownloader.download(url: url)
+            
+            XCTAssert(FileManager.default.fileExists(atPath: path))
+            
+            try? FileManager.default.removeItem(atPath: path)
+        } catch {
+            XCTFail("\(error)")
+        }
     }
 }
